handle errors for retrieving conflict data
