+++ components/examine/recipe/match/XproCorp.vue
@@ -28,7 +28,9 @@
 
       <h3 class="font-bold">Head Office</h3>
       <div>
-        <p v-for="addrLine in conflict['head office']">{{ addrLine }}</p>
+        <p v-for="addrLine in parseAddress(conflict['head office'])">
+          {{ addrLine }}
+        </p>
       </div>
     </div>
   </div>
@@ -36,6 +38,7 @@
 
 <script setup lang="ts">
 import type { XproCorporation } from '~/types'
+import { parseAddress } from '~/util'
 
 defineProps<{
   conflict: XproCorporation
